fix tensorflow cast on windows

+++ src/data.py
@@ -386,7 +386,7 @@ class AudioProcessor(object):
                                 len(video_channels)),
                                dtype=np.float32)
         bkg = {}
-      labels = np.zeros(nsounds, dtype=int)
+      labels = np.zeros(nsounds, dtype=np.int32)
       # repeatedly to generate the final output sound data we'll use in training.
       for i in range(offset, offset + nsounds):
         # Pick which sound to use.
